feature: upgrade mybatis plus to 3.5.1

+++ scaleph-core/scaleph-core-di/src/main/java/cn/sliew/scaleph/core/di/service/DiClusterConfigService.java
@@ -25,5 +25,5 @@ public interface DiClusterConfigService {
 
     DiClusterConfigDTO selectOne(Long id);
 
-    int totalCnt();
+    Long totalCnt();
 }
