+++ scaleph-core/scaleph-core-di/src/main/java/cn/sliew/scaleph/core/di/service/DiProjectService.java
@@ -64,7 +64,7 @@ public interface DiProjectService {
      */
     List<DiProjectDTO> listAll();
 
-    int totalCnt();
+    Long totalCnt();
 
     DiProjectDTO selectOne(Serializable id);
 }
