+++ scaleph-core/scaleph-core-di/src/main/java/cn/sliew/scaleph/core/di/service/impl/DiProjectServiceImpl.java
@@ -108,7 +108,7 @@ public class DiProjectServiceImpl implements DiProjectService {
     }
 
     @Override
-    public int totalCnt() {
+    public Long totalCnt() {
         return this.diProjectMapper.selectCount(null);
     }
